Fixed Mac-compatibility, added Mac framework, added Mac Unit Tests removed obsolte methods

+++ Core/Source/BCKEAN13Code.m
@@ -78,7 +78,7 @@ static char *variant_patterns[10] = {"LLLLLLRRRRRR",  // 0
 		{
 			if (error)
 			{
-				NSString *message = [NSString stringWithFormat:@"%@ cannot encode '%@' at index %d", NSStringFromClass([self class]), character, index];
+				NSString *message = [NSString stringWithFormat:@"%@ cannot encode '%@' at index %d", NSStringFromClass([self class]), character, (int)index];
 				*error = [NSError BCKCodeErrorWithMessage:message];
 			}
