+++ Core/Source/BCKUPCACode.m
@@ -48,7 +48,7 @@
 		{
 			if (error)
 			{
-				NSString *message = [NSString stringWithFormat:@"%@ cannot encode '%@' at index %d", NSStringFromClass([self class]), character, index];
+				NSString *message = [NSString stringWithFormat:@"%@ cannot encode '%@' at index %d", NSStringFromClass([self class]), character, (int)index];
 				*error = [NSError BCKCodeErrorWithMessage:message];
 			}
